third_party/blink/web_tests/http/tests/security/mask-image-cors-port-expected.txt

CONSOLE ERROR: Access to image at 'http://localhost:8080/resources/square100.png' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
